chore: refactoring the completion support
