+++ src/documentdb/query-language/shared/completionKnowledge.ts
