+++ src/documentdb/query-language/shared/cursorContext.ts
