+++ src/documentdb/query-language/shared/extractQuotedKey.ts
