+++ src/documentdb/query-language/shared/index.ts
