+++ src/documentdb/query-language/shared/jsGlobalDefs.ts
