+++ src/documentdb/query-language/shared/snippetUtils.ts
