+++ src/documentdb/query-language/shared/tdd/sharedCompletionLogic.test.ts
