+++ src/documentdb/query-language/shared/typeSuggestionData.ts
