+++ src/webviews/documentdbQuery/documentdbQueryCompletionProvider.test.ts
@@ -11,6 +11,7 @@ import {
 } from '@vscode-documentdb/documentdb-constants';
 // eslint-disable-next-line import/no-internal-modules
 import type * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api';
+import { type CursorContext } from '../../documentdb/query-language/shared/cursorContext';
 import { clearAllCompletionContexts, setCompletionContext } from './completionStore';
 import {
     createCompletionItems,
@@ -25,7 +26,6 @@ import {
     stripOuterBraces,
 } from './documentdbQueryCompletionProvider';
 import { EditorType } from './languageConfig';
-import { type CursorContext } from './shared/cursorContext';
 
 /**
  * Minimal mock of `monaco.languages.CompletionItemKind` for testing.
