+++ src/webviews/documentdbQuery/registerLanguage.ts
@@ -26,13 +26,13 @@
 
 // eslint-disable-next-line import/no-internal-modules
 import type * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api';
+import { detectCursorContext } from '../../documentdb/query-language/shared/cursorContext';
+import { extractQuotedKey } from '../../documentdb/query-language/shared/extractQuotedKey';
 import { getCompletionContext } from './completionStore';
 import { createCompletionItems } from './documentdbQueryCompletionProvider';
 import { getHoverContent } from './documentdbQueryHoverProvider';
 import { isCursorInsideString } from './isCursorInsideString';
 import { LANGUAGE_ID, parseEditorUri } from './languageConfig';
-import { detectCursorContext } from './shared/cursorContext';
-import { extractQuotedKey } from './shared/extractQuotedKey';
 
 /** Coalesces concurrent registrations into a single promise. */
 let registrationPromise: Promise<void> | undefined;
