+++ src/webviews/documentdbQuery/tdd/completionBehavior.test.ts
@@ -26,10 +26,10 @@
 
 // eslint-disable-next-line import/no-internal-modules
 import type * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api';
+import { type CursorContext } from '../../../documentdb/query-language/shared/cursorContext';
 import { clearAllCompletionContexts, setCompletionContext } from '../completionStore';
 import { createCompletionItems } from '../documentdbQueryCompletionProvider';
 import { EditorType } from '../languageConfig';
-import { type CursorContext } from '../shared/cursorContext';
 
 // ---------- Test infrastructure ----------
